extracted downloadables for page (no styling yet)
